This is a first commit to a new branch, At this stage UserFundsTransactions Work and i can add funds Or subtract funds from a "Wallet" basically. There is a bit of a problem that currently basically user can have multiple funds accounts, which is not bad per say, but that can be currenly created only at the backend and adding to this wallet is kind of janky which is a bit stupid. Its also not necessarily a thing i want to get rid of because user could potentially want to make separate accounts lets say for separate currencies and that makes sense in a way. Currently i leave it be and go on with some basic functionality.
At this stage the link between a wallet and buying an asset is not there yet, so it would be wise to work on that for the time. Because that would kind of round up basic functionality in its entirety. After that major missing components would be creating an user and perhaps reports or a portfolio view which should give overall status to the user at a glance.